+++ CHANGES.md
@@ -1,4 +1,18 @@
-0.0
+0.1
 ---
 
-- Initial version.
+- Initial Version
+ - Restful view
+ - Google geolocation interface
+- Caching
+ - Added DictBackend
+ - Decorator
+- Dockerised
+ - Added Entry Script
+ - Setup the Dockerfile
+- Documentation
+ - Added Readme.md
+ - Documented inline code
+
+
+